update hybrid to new api
